Upgrade NPM dependencies: webpack, tailwindcss, highlightjs
Upgrade to webpack v5 with necessary webpack.config.js changes and Ruby tag helper
changes

Upgrade to tailwindcss v2

Upgrade to highlightjs v11 with api changes

Upgrade to latest babel, eslint
