+++ package.json
@@ -7,8 +7,9 @@
     "doc": "doc"
   },
   "scripts": {
-    "build:prod": "NODE_ENV=production webpack --bail -p",
-    "build:dev": "NODE_ENV=development webpack --watch -d --progress --color",
+    "build:prod": "NODE_ENV=production webpack --bail --progress",
+    "build:dev": "NODE_ENV=development webpack --watch --progress --color",
+    "configtest": "webpack configtest",
     "lint": "eslint webpack.config.js config source/assets/javascripts",
     "test": "echo \"No test specified\" && exit 0"
   },
@@ -20,42 +21,42 @@
   },
   "homepage": "https://github.com/rossta/rossta.github.com#readme",
   "dependencies": {
-    "@babel/core": "^7.6.4",
+    "@babel/core": "^7.20.12",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/preset-env": "^7.6.3",
-    "@fullhuman/postcss-purgecss": "^1.3.0",
-    "autoprefixer": "^9.7.0",
+    "@babel/preset-env": "^7.20.2",
+    "@fullhuman/postcss-purgecss": "^5.0.0",
+    "autoprefixer": "^10.4.13",
     "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.0.6",
-    "babel-plugin-module-resolver": "^3.0",
+    "babel-loader": "^9.1.2",
+    "babel-plugin-module-resolver": "^5.0.0",
     "babel-polyfill": "^6.26",
-    "clean-webpack-plugin": "^3.0.0",
-    "compression-webpack-plugin": "^3.0.0",
-    "css-loader": "^3.2.0",
-    "debug": "^4.1.1",
-    "eslint": "^7.6.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "exports-loader": "^0.7.0",
-    "expose-loader": "^0.7",
-    "highlight.js": "^9.16.2",
-    "imports-loader": "^0.8.0",
-    "mini-css-extract-plugin": "^0.8.0",
-    "node-sass": "^7.0.0",
-    "postcss": "^7.0.23",
-    "postcss-import": "^12.0.1",
-    "postcss-loader": "^3.0.0",
-    "prettier": "^2.0.5",
-    "sass-loader": "^8.0.0",
-    "serialize-javascript": "^3.0.0",
-    "style-loader": "^1.0.0",
-    "tailwindcss": "^1.1.3",
-    "terser-webpack-plugin": "^2.2.1",
-    "webpack": "4.41.2",
-    "webpack-bundle-analyzer": "^3.6.0",
-    "webpack-cli": "^3.3.10",
-    "webpack-manifest-plugin": "^2.2.0"
+    "clean-webpack-plugin": "^4.0.0",
+    "compression-webpack-plugin": "^10.0.0",
+    "css-loader": "^6.7.3",
+    "debug": "^4.3.4",
+    "eslint": "^8.33.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-prettier": "^4.2.1",
+    "exports-loader": "^4.0.0",
+    "expose-loader": "^4.0.0",
+    "highlight.js": "^11.7.0",
+    "imports-loader": "^4.0.1",
+    "mini-css-extract-plugin": "^2.7.2",
+    "node-sass": "^8.0.0",
+    "postcss": "^8.4.21",
+    "postcss-import": "^15.1.0",
+    "postcss-loader": "^7.0.2",
+    "prettier": "^2.8.4",
+    "sass-loader": "^13.2.0",
+    "serialize-javascript": "^6.0.1",
+    "style-loader": "^3.3.1",
+    "tailwindcss": "^2.2.19",
+    "terser-webpack-plugin": "^5.3.6",
+    "webpack": "5.75.0",
+    "webpack-assets-manifest": "^5.1.0",
+    "webpack-bundle-analyzer": "^4.7.0",
+    "webpack-cli": "^5.0.1"
   },
   "resolutions": {
     "*/serialize-javascript": "^2.1.1"
