+++ source/layouts/layout.erb
@@ -26,7 +26,7 @@
       <%= yield %>
     </main>
     <%= partial "partials/footer" %>
-    <%= javascript_pack_tag "runtime", "vendors~app", "app" %>
+    <%= javascript_pack_tag "app" %>
     <%= yield_content :javascripts %>
   </body>
 </html>
